Script for solving Einstein's Zebra Puzzle Game
It can be easily solved by hand.
Since the book I read containing this puzzle describing vaguely about whether "Green is to the right of White" or "Green is immediate right of White", so there is a boolean variable in the code to decide which condition to take.

Note : pure burteforce will take 120^5, which is unaffordable

Tips on solving this puzzle:
if "Green is to the right of White", the constraint is too weak, nearly random filling in the blank will do.
if "Green is to the immediate right of White", then we can quickly find 1st house on the Left is yellow, and the only branch is to decide whether 3rd is red or 5th is red. then simply fill in the blank.

Question remains: how to construct such a puzzle?